feat: implements prism  and CRUD

+++ src/modules/contacts/contacts.service.ts
@@ -1,29 +1,79 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateContactDto } from './dto/create-contact.dto';
 import { UpdateContactDto } from './dto/update-contact.dto';
 import { ContactsRepository } from './repositories/contact.repository';
-import { contacts } from 'src/database/db';
 
 @Injectable()
 export class ContactsService {
   constructor(private contactsRepository: ContactsRepository) {}
-  create(createContactDto: CreateContactDto) {
-    return this.contactsRepository.create(createContactDto);
+  async create(createContactDto: CreateContactDto) {
+    // const user_id = await this.contactsRepository.findOne(
+    //   createContactDto.user_id,
+    // );
+    // console.log(user_id);
+    // if (!user_id) {
+    //   throw new NotFoundException('User not found');
+    // }
+    try {
+      return await this.contactsRepository.create(createContactDto);
+    } catch (error) {
+      if (createContactDto.email) {
+        throw new ConflictException('E-mail already registered');
+      }
+      throw error;
+    }
   }
 
-  findAll() {
-    return contacts;
+  async findByEmail(email: string) {
+    const contact = await this.contactsRepository.findByEmail(email);
+    return contact;
   }
 
-  // findOne(id: string) {
-  //   return this.contactsRepository.
+  // async findOneUserId(user_id: string) {
+  //   const contact = await this.contactsRepository.findUserId(user_id);
+  //   if (!contact) {
+  //     throw new NotFoundException('Contact not Found');
+  //   }
+  //   return contact;
   // }
 
-  update(id: string, updateContactDto: UpdateContactDto) {
-    return this.contactsRepository.update(id, updateContactDto);
+  async findOne(id: string) {
+    const contact = await this.contactsRepository.findOne(id);
+    if (!contact) {
+      throw new NotFoundException('Contact not Found');
+    }
+    return contact;
+  }
+  async findAll() {
+    return;
   }
 
-  remove(id: string) {
-    return this.contactsRepository.delete(id);
+  async update(id: string, updateContactDto: UpdateContactDto) {
+    const contact = await this.contactsRepository.findOne(id);
+
+    if (!contact) {
+      throw new NotFoundException('contact not found');
+    }
+
+    try {
+      return await this.contactsRepository.update(id, updateContactDto);
+    } catch (error) {
+      if (updateContactDto.email) {
+        throw new ConflictException('Contact already exists');
+      }
+      throw error;
+    }
+  }
+
+  async remove(id: string) {
+    const contactId = await this.contactsRepository.findOne(id);
+    if (!contactId) {
+      throw new NotFoundException('Contact not found');
+    }
+    return await this.contactsRepository.delete(id);
   }
 }
